pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends"

[project]
name = "routes-monitor"
version = "0.1.0"
description = "Traffic monitoring tool using Google Routes API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
routes-monitor = "routes_monitor.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
